+++ mayan/apps/mailer/models.py
@@ -5,12 +5,13 @@ import logging
 
 from django.contrib.sites.models import Site
 from django.core import mail
-from django.db import models
+from django.db import models, transaction
 from django.template import Context, Template
 from django.utils.html import strip_tags
 from django.utils.module_loading import import_string
 from django.utils.translation import ugettext_lazy as _
 
+from .events import event_email_sent
 from .managers import UserMailerManager
 from .utils import split_recipient_list
 
@@ -125,7 +126,7 @@ class UserMailer(models.Model):
 
         return super(UserMailer, self).save(*args, **kwargs)
 
-    def send(self, to, subject='', body='', attachments=None):
+    def send(self, to, subject='', body='', attachments=None, _event_action_object=None, _user=None):
         """
         Send a simple email. There is no document or template knowledge.
         attachments is a list of dictionaries with the keys:
@@ -150,14 +151,19 @@ class UserMailer(models.Model):
 
             email_message.attach_alternative(body, 'text/html')
 
+        with transaction.atomic():
             try:
                 email_message.send()
             except Exception as exception:
                 self.error_log.create(message=exception)
             else:
                 self.error_log.all().delete()
+                event_email_sent.commit(
+                    actor=_user, action_object=_event_action_object,
+                    target=self
+                )
 
-    def send_document(self, document, to, subject='', body='', as_attachment=False):
+    def send_document(self, document, to, subject='', body='', as_attachment=False, _user=None):
         """
         Send a document using this user mailing profile.
         """
@@ -190,7 +196,8 @@ class UserMailer(models.Model):
 
         return self.send(
             attachments=attachments, body=body_html_content,
-            subject=subject_text, to=to,
+            subject=subject_text, to=to, _event_action_object=document,
+            _user=_user
         )
 
     def test(self, to):